Add DISTINCT to LoadFilters query, Scan() to evmtypes
Using DISTINCT keyword in the LoadFilters query removes the need
for de-duping in Scan().

Without the deduping workaround, AddressArray and HashArray types
become more generally useful, so moving them into evm/types/types.go

Also:
  - Updated comments
  - Fixed logic around updating filter message, and changed level from
    Error to Warn.  Adding new contracts or topics to an existing filter
    might be a useful feature at some point, but since we don't use it...
    right now it's most likely an indication of a filter naming conflict,
    but doesn't cause any immediate problem since they just get merged together.
    The problem would come later: if one of the jobs gets deleted then both
    would stop working.
